+++ tests/setup_1_basic.nim
@@ -1,0 +1,8 @@
+import ../commandant
+
+commandline:
+      argument(integer, int)
+      flag(flag, "flag", "f")
+      option(option, float, "option", "o")
+
+echo integer, " ", flag, " ", option