+++ tests/setup_6_spaces.nim
@@ -1,0 +1,7 @@
+import ../commandant
+
+commandline:
+  argument first_required, string
+  option first_optional, string, "optional1", "o1"
+
+echo first_required , " " , first_optional