+++ tests/setup_7_multiple.nim
@@ -1,0 +1,7 @@
+import ../commandant
+
+commandline:
+  arguments(expendables, int, false)
+  argument required, string
+
+echo expendables, " ", required