+++ tests/setup_8_optionalMultiple.nim
@@ -1,0 +1,5 @@
+import ../commandant
+commandline:
+  arguments(expendables, int, false)
+
+echo len(expendables)